chore: properly init rbac-rules in example compose
